Fix matchNodeToChild special case with LogicOr
Previously we had only expected to enter nodes that already designated as parent/child in the tree inside the matchNodeToChild check, so we automatically returned true in the case of isinstance(parent, Group) and isintance(child, LogicOr).

This fix makes it so that we at least check that child.parent is the parent node.